+++ mutwo/abjad_utilities/__init__.py
@@ -1,7 +1,8 @@
 from .tests import *
 
 from . import tests
+from . import tools  # core patch
 
 __all__ = tests.__all__
 
-del tests
+del tests, tools